remove scipy from cncmark
