+++ .github/workflows/ci.yml
@@ -65,23 +65,3 @@ jobs:
 
       - name: Run the automated tests
         run: poetry run pytest -v
-
-  ui:
-    strategy:
-      matrix:
-        node-version: [18.x]
-        os: [ubuntu-latest]
-
-    runs-on: ${{ matrix.os }}
-    defaults:
-      run:
-        working-directory: ui
-    steps:
-      - uses: actions/checkout@v3
-      - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v3
-        with:
-          node-version: ${{ matrix.node-version }}
-      - run: yarn install
-      - run: yarn lint
-      - run: yarn build
