+++ cncmark/pyproject.toml
@@ -1,15 +1,14 @@
 [tool.poetry]
 name = "cncmark"
-version = "0.2.2"
+version = "0.2.4"
 description = ""
 authors = ["yuichiroaoki <45054071+yuichiroaoki@users.noreply.github.com>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "3.9.17"
+python = ">=3.9.2, <3.10"
 numpy-stl = "^3.0.1"
 mysql-connector-python = "^8.1.0"
-scipy = "^1.11.1"
 
 
 [tool.poetry.group.dev.dependencies]
